Graphics.UI.FLTK.LowLevel.Dispatch

FindOp

type family FindOp orig hierarchy needle :: *

Equations

FindOp orig () n = NoFunction n orig

FindOp orig hierarchy needle = FindOpHelper orig hierarchy needle (Contains (Functions hierarchy) needle)

Functions

type family Functions x :: *

Associate a "class" with it's member functions

Match

data Match a

See FindOp for more details.

NoFunction

data NoFunction a b

See FindOp for more details.

Op

class Op op obj origObj impl where

Implementations of methods on various types of objects.

op - name of the function

obj - the class that implements op

origObj - the class in the hierarchy where the search for op started. Kept around in case the type in needed. The best example is setCallback whose implementation is usually found much lower in the hierarchy but where we also want to enforce that the implementation take the type of the widget calling it.

impl - a function that takes the a Ref origobj, casted down to Ref obj and whatever other parameters the instance specifies.

Methods

runOp :: op -> origObj -> Ref obj -> impl

dispatch

dispatch :: forall op obj origObj impl. (Match obj ~ FindOp origObj origObj op, Op op obj origObj impl) => op -> Ref origObj -> impl

Given some member function op and a Ref to some class origObj return the implementation of op. See Op for more details.

Every FLTK function called on some Ref uses this function to figure out what arguments it needs.

runOp

castTo

castTo :: Ref a -> Ref r

Cast any reference to any other reference. Unsafe, intended to be used by Op.

safeCast

safeCast :: Parent a r => Ref a -> Ref r

Cast any reference to one of its ancestors.

Parent

class Parent a b

A class with a single instance that is found only if b is an ancestor of a.

Used by some Op implementations to enforce that certain parameters have to be at least a b.